refactor: reorganize commands
