+++ src/commands.ts
@@ -3,33 +3,33 @@ import * as vscode from 'vscode';
 // Local imports - commands
 import { registerFileSelectionCommands } from '@commands/files/fileSelectionCommands';
 import { registerLatexdiffCommands } from '@commands/latex/latexdiffCommands';
-import { registerGitCommands } from '@commands/git/gitCommands';
-import { registerPackCommands } from './commands/packCommands';
-import { registerCleanCommands } from './commands/cleanCommands';
-import { registerMergeCommands } from './commands/mergeCommands';
-import { registerExecuteCommand } from './commands/executeCommand';
+import { registerGitCommands } from '@commands/integrations/git/gitCommands';
+import { registerPackCommands } from './commands/workspace/packCommands';
+import { registerCleanCommands } from './commands/workspace/cleanCommands';
+import { registerMergeCommands } from './commands/workspace/mergeCommands';
+import { registerExecuteCommand } from './commands/agent/executeCommand';
 import { TeXRAViewProvider } from './ViewProvider';
 import { registerLatexCommands } from '@commands/latex/latexCommands';
-import { registerImageCommands } from './commands/imageCommands';
-import { registerFigureCommands } from './commands/figCommands';
-import { registerTestCommands } from './commands/testCommands';
-import { registerXmlCommands } from './commands/xmlCommands';
-import { registerYamlCommands } from './commands/yamlCommands';
-import { registerAgentCommands } from './commands/agentCommands';
-import { registerAgentCreatorCommands } from '@commands/agentCreatorCommands';
-import { registerApiKeyCommands } from '@commands/api/apiKeyCommands';
-import { registerStateRestoreCommand } from './commands/stateRestoreCommand';
-import { registerTextEditorCommands } from './commands/textEditorCommands';
+import { registerImageCommands } from './commands/latex/imageCommands';
+import { registerFigureCommands } from './commands/latex/figCommands';
+import { registerTestCommands } from './commands/tests/testCommands';
+import { registerXmlCommands } from './commands/workspace/xmlCommands';
+import { registerYamlCommands } from './commands/workspace/yamlCommands';
+import { registerAgentCommands } from './commands/agent/agentCommands';
+import { registerAgentCreatorCommands } from '@commands/agent/agentCreatorCommands';
+import { registerApiKeyCommands } from '@commands/integrations/apiKeyCommands';
+import { registerStateRestoreCommand } from './commands/utils/stateRestoreCommand';
+import { registerTextEditorCommands } from './commands/utils/textEditorCommands';
 import { registerLinterCommands } from '@commands/latex/linterCommands';
-import { registerWolframAlphaCommands } from '@commands/wolfram/wolframAlphaCommands';
-import { registerWolframScriptCommands } from '@commands/wolfram/wolframScriptCommands';
-import { registerHistoryCommands } from '@commands/history/historyCommands';
-import { registerArXivCommands } from './commands/arXivCommands';
-import { registerCompareCommands } from './commands/compareCommands';
-import { registerHelpCommands } from './commands/helpCommands';
-import { registerProgressViewCommands } from '@commands/progress/progressViewCommands';
+import { registerWolframAlphaCommands } from '@commands/integrations/wolfram/wolframAlphaCommands';
+import { registerWolframScriptCommands } from '@commands/integrations/wolfram/wolframScriptCommands';
+import { registerHistoryCommands } from '@commands/views/history/historyCommands';
+import { registerArXivCommands } from './commands/integrations/arXivCommands';
+import { registerCompareCommands } from './commands/utils/compareCommands';
+import { registerHelpCommands } from './commands/utils/helpCommands';
+import { registerProgressViewCommands } from '@commands/views/progressViewCommands';
 import { registerOpenFileCommands } from '@commands/files/openFileCommands';
-import { registerSettingsCommands } from './commands/settingsCommands';
+import { registerSettingsCommands } from './commands/utils/settingsCommands';
 
 import * as logger from '@logger/logUtils';
 
@@ -92,26 +92,26 @@ export function registerCommands(context: vscode.ExtensionContext) {
 // Add exports for the command modules
 export { fileSelectionCommands } from '@commands/files/fileSelectionCommands';
 export { latexdiffCommands } from '@commands/latex/latexdiffCommands';
-export { gitCommands } from '@commands/git/gitCommands';
-export { packCommands } from './commands/packCommands';
-export { mergeCommands } from './commands/mergeCommands';
-export { executeCommand } from './commands/executeCommand';
+export { gitCommands } from '@commands/integrations/git/gitCommands';
+export { packCommands } from './commands/workspace/packCommands';
+export { mergeCommands } from './commands/workspace/mergeCommands';
+export { executeCommand } from './commands/agent/executeCommand';
 export { latexCommands } from '@commands/latex/latexCommands';
-export { imageCommands } from './commands/imageCommands';
-export { figureCommands } from './commands/figCommands';
-export { testCommands } from './commands/testCommands';
-export { xmlCommands } from './commands/xmlCommands';
-export { yamlCommands } from './commands/yamlCommands';
-export { apiKeyCommands } from '@commands/api/apiKeyCommands';
-export { stateRestoreCommand } from './commands/stateRestoreCommand';
-export { textEditorCommands } from './commands/textEditorCommands';
+export { imageCommands } from './commands/latex/imageCommands';
+export { figureCommands } from './commands/latex/figCommands';
+export { testCommands } from './commands/tests/testCommands';
+export { xmlCommands } from './commands/workspace/xmlCommands';
+export { yamlCommands } from './commands/workspace/yamlCommands';
+export { apiKeyCommands } from '@commands/integrations/apiKeyCommands';
+export { stateRestoreCommand } from './commands/utils/stateRestoreCommand';
+export { textEditorCommands } from './commands/utils/textEditorCommands';
 export { linterCommands } from '@commands/latex/linterCommands';
-export { wolframAlphaCommands } from '@commands/wolfram/wolframAlphaCommands';
-export { wolframScriptCommands } from '@commands/wolfram/wolframScriptCommands';
-export { historyCommands } from '@commands/history/historyCommands';
-export { arXivCommands } from './commands/arXivCommands';
-export { compareCommands } from './commands/compareCommands';
-export { helpCommands } from './commands/helpCommands';
-export { agentCreatorCommands } from '@commands/agentCreatorCommands';
+export { wolframAlphaCommands } from '@commands/integrations/wolfram/wolframAlphaCommands';
+export { wolframScriptCommands } from '@commands/integrations/wolfram/wolframScriptCommands';
+export { historyCommands } from '@commands/views/history/historyCommands';
+export { arXivCommands } from './commands/integrations/arXivCommands';
+export { compareCommands } from './commands/utils/compareCommands';
+export { helpCommands } from './commands/utils/helpCommands';
+export { agentCreatorCommands } from '@commands/agent/agentCreatorCommands';
 export { registerOpenFileCommands as openFileCommands } from '@commands/files/openFileCommands';
-export { settingsCommands } from './commands/settingsCommands';
+export { settingsCommands } from './commands/utils/settingsCommands';
