+++ src/commands/agent/executeCommand.ts
