+++ src/commands/integrations/wolfram/wolframScriptCommands.ts
