+++ src/commands/latex/figCommands.ts
