+++ src/commands/latex/imageCommands.ts
