+++ src/commands/utils/compareCommands.ts
