+++ src/commands/utils/textEditorCommands.ts
