+++ src/commands/views/progressViewCommands.ts
